Using ArchiveConfig directly in Archive
Instead of having an ArchiveConfig model in code that's serialized/deserialized into a dict to enter DB (stored as JSONB)
We now declare the Archive.config as a proper ArchiveConfig and have the serialization done automatically.
This preserves type hints all the way and allows properties access

ArchiveConfig now moved to model as it's part (sub) of a model.

In addition, ArchiveConfig gets an `is_ready()` method that checks values with scraperlib preventing request of incorrect archive config