Fixed #95: Allow Single User
This feature serves two purposes:

- sets a variable on both frontend and backend indicating that we are in single-user mode. this is the concept we'll use to switch between public SaaS and private single-user
- sets a single User ID that is sent to all ~users~

This makes sure that anyone connecting is on the same session, sees the same projects. Even though the cookie might expire, re-requesting one gets the same User ID.

On the backend, if SINGLE_USER_ID is set, it is checked for existence on start or created if not.
In the frontend, homepage checks whether it is expecting a single user mode (ENV) and if yes, fetches the User and Projects from backend before redirecting to /collections